Desktop Video 10.8.6 SDK (13 Apr 2017) This SDK provides developer support for Desktop Video 10.8.6 that allows updating of hardware control and software interfaces for Desktop Video products.
Desktop Video 10.8.6 (13 Apr 2017)
This software update adds support for SDI output slew rate support for Developers on DeckLink Duo 2 and DeckLink Quad 2, which allows developers to accurately synchronise the output of playback to remote locations that are not on the same genlock environment. This feature can be accessed via the DeckLink API. This update also adds compatibility improvements for Media Composer plus general performance improvements for all DeckLink, UltraStudio and Intensity models.